Create spring security config and basic schema for project

+++ src/main/java/com/pajelonek/clipwatcher/accessingdatamysql/Employee.java
@@ -5,6 +5,7 @@ import lombok.Data;
 
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @AllArgsConstructor
 @Data
@@ -15,6 +16,6 @@ public class Employee {
     private int id;
     private String name;
     private String dept;
-    private long salary;
+    private BigDecimal salary;
 
 }
